tools/seeder/pyproject.toml
[tool.poetry]
name = "seeder"
version = "0.1.0"
description = ""
authors = ["baseplate-admin <61817579+baseplate-admin@users.noreply.github.com>"]
license = "AGPL-v3-only"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
humanize = "^4.6.0"
termcolor = "^2.2.0"
requests-cache = {extras = ["redis"], version = "^1.0.0"}
requests-ratelimiter = "*"
python-dotenv = "^1.0.0"
selectolax = "*"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
poethepoet = "^0.18.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = [
  "poetry-core",
]

[tool.poe.tasks]
seed_character = 'python seed.py character'
seed_staff = 'python seed.py staff'
seed_genre = 'python seed.py genre'
seed_themes = 'python seed.py theme'
seed_producers = 'python seed.py producer'